fix: tighten publisher SHA decode, failpoints, and worker trust schemas
Decode isolated publication heads as GitCommitSha, require
IngressStoreFailpoint in the store Layer, and decode worker requests
with the canonical PublisherTrust and check schemas.
